Add test code to deploy river registry to simulated.Backend for go ... (#5375)
...testing

+++ casablanca/node/config/config.go
@@ -67,7 +67,7 @@ type DatabaseConfig struct {
 
 type ChainConfig struct {
 	NetworkUrl      string
-	ChainId         int
+	ChainId         uint64
 	FakeBlockTimeMs int64
 
 	// TODO: these need to be removed from here
